WebDriver: convert UChar to char16_t https://bugs.webkit.org/show_bug.cgi?id=294729 rdar://153822236
Reviewed by Alex Christensen.

WebKit wishes to move from UChar to the C++ standard char16_t, now this is
universally supported. This makes the change in WebDriver. This is a simple
textual substitution of \bUChar\b with char16_t. These types are currently
typedef'ed to be the same so no functional changes are expected.

* Source/WebDriver/Session.cpp:
(WebDriver::Session::virtualKeyForKey):
* Source/WebDriver/Session.h:

Canonical link: https://commits.webkit.org/296801@main