Content with a grayscale filter and mix-blend-mode don't get mix-blend-mode applied. https://bugs.webkit.org/show_bug.cgi?id=293903 <rdar://152460888>
Reviewed by Simon Fraser.

Some filter types (including grayscale) are implemented by pushing a
transparency layer, rather than diverting to a temporary context.

In those cases, the code attempting to setup a transparency layer for opacity
(and mix-blend-mode) ends up doing so too late, and applies the mix-blend-mode
inside the filter (where it has nothing to blend with).

Move this up to before the filter is setup, so that the ordering is always
correct.

* LayoutTests/imported/w3c/web-platform-tests/css/filter-effects/filter-with-mix-blend-mode-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/filter-effects/filter-with-mix-blend-mode.html: Added.
* Source/WebCore/rendering/RenderLayer.cpp:
(WebCore::RenderLayer::paintLayerContents):

Canonical link: https://commits.webkit.org/296827@main