AX: Add a failsafe to the while loop in AXCoreObject::partialOrder https://bugs.webkit.org/show_bug.cgi?id=295222 rdar://154696128
Reviewed by Joshua Hoffman.

When testing a different bug, I was able to enter an infinite loop in this function. I tried to rebuild with logging
to debug, but wasn't able to hit it again despite much effort.

To avoid a permanent hang for our users, add a failsafe counter, and an ASSERT which should hopefully help us debug the
root cause when it triggers.

* Source/WebCore/accessibility/AXCoreObject.cpp:
(WebCore::AXCoreObject::partialOrder):

Canonical link: https://commits.webkit.org/296846@main