Reduce unsafeness in highlight module again https://bugs.webkit.org/show_bug.cgi?id=295264
Reviewed by Ryosuke Niwa.

The initial attempt at applying
https://github.com/WebKit/WebKit/wiki/Safer-CPP-Guidelines in
296813@main was backed out in 296852@main due to Highlight API tests
timing out.

The root cause of this is 290446@main which introduced
Document::protectedAppHighlightRegistry but made it wrap
Document::highlightRegistry rather than Document::appHighlightRegistry.
That is an incorrect functional change as
Document::appHighlightRegistry is what was used.

A second usage of Document::protectedAppHighlightRegistry was
introduced in 290504@main again in
Source/WebKit/WebProcess/WebPage/WebPage.cpp and there too it was a
functional change as Document::appHighlightRegistry is what was used.

So now we fix Document::protectedAppHighlightRegistry by making it wrap
Document::appHighlightRegistry as was (presumably) originally intended.
This constitues a regression fix for the two existing usages in
WebPage.cpp, although thus far they went unnoticed.

Canonical link: https://commits.webkit.org/296869@main