[Model Element] Model with autoplay sometimes does not play after page reload https://bugs.webkit.org/show_bug.cgi?id=295221 rdar://154646578
Reviewed by Mike Wyrzykowski.

This issue happens under a timing condition where the model got suspended
before the model process has finished loading it. That can happen if the model
process crashes during model load, or if page navigation happens before the
model finishes loading.

In HTMLModelElement::unloadModelPlayer(), we try to save off the current state
of the model player before the deleting the model player, so we can restore to
that state when the element is resumed and a new model player is created.
However, if the model player hasn't even finished loading when unloadModelPlayer()
is called, its animation/transform states won't be valid. Before this fix, the
invalid animation state returned has paused=true, so after the model element is
resumed and the model player is created with that animation state, it then
incorrectly pauses the animation.

To fix this, return null for ModelProcessModelPlayer::currentAnimationState()
and ModelProcessModelPlayer::currentTransformState() if the model load has not
completed successfully. We can tell whether the model load has finished loading
by checking for the presence of a non-null m_boundingBoxExtents.

* LayoutTests/model-element/model-element-suspend-before-ready-expected.txt: Added.
* LayoutTests/model-element/model-element-suspend-before-ready.html: Added.
* Source/WebKit/WebProcess/Model/ModelProcessModelPlayer.cpp:
(WebKit::ModelProcessModelPlayer::currentAnimationState const):
(WebKit::ModelProcessModelPlayer::currentTransformState const):

Canonical link: https://commits.webkit.org/296870@main