[WebGPU] Move GPUAdapter isFallbackAdapter attribute to GPUAdapterInfo https://bugs.webkit.org/show_bug.cgi?id=290362 rdar://147800092
Reviewed by Tadeu Zagallo.

The spec deprecated this in March and is removing it
from GPUAdapter, but we don't want to break any clients,
so just add it in the new place for now.

We never create a fallback adapter, so we can always
return false.

* Source/WebCore/Modules/WebGPU/GPUAdapterInfo.h:
(WebCore::GPUAdapterInfo::isFallbackAdapter const):
* Source/WebCore/Modules/WebGPU/GPUAdapterInfo.idl:

Canonical link: https://commits.webkit.org/296879@main