object-fit: scale-down is ignored when srcset attribute and w descriptor is used https://bugs.webkit.org/show_bug.cgi?id=295203 rdar://154687132
Reviewed by Simon Fraser.

This patch aligns WebKit with Blink / Chromium.

Merge: https://chromium.googlesource.com/chromium/src.git/+/80441f0a1eb089bf4e9d755a7af5c3bfd3b787ec

The object-fit: scale-down property requires that the image be
sized to the smaller of "none" or "contain". When a srcset is used,
the intrinsicSize of the image is scaled such that the srcset fills
the expected area. This causes scale-down to consider the intrinsic
size to be bigger than the actual image, and size things too large.

This change modifies the intrinsic size for content-fit: scale-down
in RenderReplaced::replacedContentRect to un-apply the
imageDevicePixelRatio, that is the scale that resized the src.

* Source/WebCore/rendering/RenderReplaced.cpp:
(WebCore::RenderReplaced::computeAspectRatioInformationForRenderBox):
(WebCore::RenderReplaced::replacedContentRect):
* LayoutTests/fast/hidpi/image-srcset-fit-scale-down.html:
* LayoutTests/fast/hidpi/image-srcset-fit-scale-down-expected.html:

Canonical link: https://commits.webkit.org/296914@main