Occasional crash in WKMouseDeviceObserver.connectedDeviceCount.setter due to precondition violation https://bugs.webkit.org/show_bug.cgi?id=295342 rdar://154860259
Reviewed by Aditya Keerthi.

This crash happens due to a violation of the precondition asserting that the number of connected devices
is >= 0. This is a perfectly valid precondition, since it doesn't make sense that there could be a _negative_
amount of connected devices.

This implies that there is some underlying issue in how the notifications are published. To workaround this
and be more resliant, access the array of connected devices directly and use its count, instead of manually
incrementing and decrementing.

* Source/WebKit/UIProcess/WKMouseDeviceObserver.swift:
(WKMouseDeviceObserver.start):

Canonical link: https://commits.webkit.org/296926@main