Adopt more UniqueRef in clipboard, webdatabase, and page https://bugs.webkit.org/show_bug.cgi?id=295393
Reviewed by Youenn Fablet.

The many changes are mainly due to the removal of
checkedFocusController() which is unneeded now m_focusController is
const UniqueRef and focusController() continues to be inline.

When focusController() is followed by a trivial method, such as
focusedOrMainFrame(), there is no need for any trivial methods in the
chain to be checked or protected. Hence we replace
protected[Core]Page() with [core]page() here and there.

Sometimes it is followed by a non-trivial method. In those cases we
ensure that whatever invokes focusController() is protected so we don't
reduce coverage.

Helps with https://github.com/WebKit/WebKit/wiki/Safer-CPP-Guidelines

Canonical link: https://commits.webkit.org/296972@main