REGRESSION (294690@main ): [ macOS wk2 ] http/wpt/service-workers/navigate-iframes-window-client.https.html is a consistent failure rdar://152266507 https://bugs.webkit.org/show_bug.cgi?id=293761
Reviewed by Pascoe.

Before 294690@main, we were always expecting a listener when processing the WebProcess message answer triggered by WebFrameProxy::navigateServiceWorkerClient.
This is no longer the case, and this check is anyway no longer needed.
We remove it as either the listener will call the navigate callback or the callback will be called when load finishes or fails.

* LayoutTests/platform/ios/TestExpectations:
* LayoutTests/platform/mac-wk2/TestExpectations:
* Source/WebKit/UIProcess/WebFrameProxy.cpp:
(WebKit::WebFrameProxy::navigateServiceWorkerClient):

Canonical link: https://commits.webkit.org/296976@main