REGRESSION (290813@main): Some videos on Amazon Prime have no audio https://bugs.webkit.org/show_bug.cgi?id=295385 rdar://154418594
Reviewed by Jean-Yves Avenard.

In 290813@main, an AVAssetTrack was always added to AVTrackPrivateAVFObjCImpl when that object was
created with an AVMediaSelectionOption. But because a number of the functions in
AVTrackPrivateAVFObjCImpl check first for the presence of an AVAssetTrack, this caused a change in
behavior where those AudioTracks created with an AVMediaSelectionOption no longer had a label.

Amazon Prime expected its AudioTracks to have labels that matched the names in its HLS
manifest, and would disable tracks not matching the label that corresponds to the audio language
selected by the user.

Resolved this in AVTrackPrivateAVFObjCImpl::label() by falling back to the AVMediaSelectionOption
when an AVAssetTrack does not have commonMetadata necessary for determining a label.

Modified http/tests/media/hls/hls-audio-tracks-language.html to assert each track's label as well
as its language.

* LayoutTests/http/tests/media/hls/hls-audio-tracks-language-expected.txt:
* LayoutTests/http/tests/media/hls/hls-audio-tracks-language.html:
* Source/WebCore/platform/graphics/avfoundation/AVTrackPrivateAVFObjCImpl.mm:
(WebCore::AVTrackPrivateAVFObjCImpl::label const):

Canonical link: https://commits.webkit.org/296981@main