[HDR] Use ExtendedSRGB instead of ExtendedDisplayP3 as the colorSpace for the layer backing stores on iOS https://bugs.webkit.org/show_bug.cgi?id=295379 rdar://154803457
Reviewed by Simon Fraser.

In 296814@main, ExtendedDisplayP3 replaced ExtendedRec2020 as the colorSpace of
the HDR layer backing stores on iOS.

It turned out that using ExtendedSRGB is more performant than using ExtendedDisplayP3.
Setting the colorSpace of the layer to ExtendedDisplayP3 requires GPU to do color
matching when compositing this layer to the window server.

* Source/WebCore/platform/graphics/ContentsFormat.cpp:
(WebCore::contentsFormatExtendedColorSpace):

Canonical link: https://commits.webkit.org/296989@main