REGRESSION(296179@main): [iOS] imported/w3c/web-platform-tests/css/css-fonts/generic-family-keywords-002.html is a constant failure https://bugs.webkit.org/show_bug.cgi?id=295422 rdar://154973303
Reviewed by Tim Nguyen.

This needs to be rebaselined. On the last WPT import sync font-family-keywords.js was updated.
This files defines the list of generic font family keywords that are tested in this test.

We still passes the same tests as before but now that are more keywords being tested.

* LayoutTests/platform/glib/imported/w3c/web-platform-tests/css/css-fonts/generic-family-keywords-002-expected.txt:
* LayoutTests/platform/gtk/TestExpectations:
* LayoutTests/platform/ios/TestExpectations:
* LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-fonts/generic-family-keywords-002-expected.txt:
* LayoutTests/platform/wpe/TestExpectations:

Canonical link: https://commits.webkit.org/297000@main