[WPE][run-benchmark] The temporary directory used as homedir is not always cleaned https://bugs.webkit.org/show_bug.cgi?id=295468
Reviewed by Adrian Perez de Castro.

The run-benchmark browser driver for Linux was not cleaning the temporary directory
used as homedir when there was an exception (like timeout) with the test run.

That is an issue on the bots because after some weeks running they end filling up
the /tmp directory (which has limited space due to being a ramdisk).

This patch modifies the base class used for the Linux browser drivers to ensure that
the temporary directory is cleaned just right after closing the browser and not only
when restoring the environment.

* Tools/Scripts/webkitpy/benchmark_runner/browser_driver/linux_browser_driver.py:
(LinuxBrowserDriver.__init__):
(LinuxBrowserDriver._clean_temp_profiledir):
(LinuxBrowserDriver.restore_env):
(LinuxBrowserDriver.close_browsers):

Canonical link: https://commits.webkit.org/297031@main