USE(GRAPHICS_CONTEXT_FILTERS) should be unconditionally enabled on PLATFORM(COCOA) https://bugs.webkit.org/show_bug.cgi?id=295508
Reviewed by Matt Woodrow.

Before 296269@main this directive was guarded as follows:

  PLATFORM(COCOA) && (HAVE(CGSTYLE_CREATE_SHADOW2) || HAVE(CGSTYLE_COLORMATRIX_BLUR))

In that commit we removed HAVE(CGSTYLE_CREATE_SHADOW2), but then
incorrectly gated this directive as follows:

  PLATFORM(COCOA) && HAVE(CGSTYLE_COLORMATRIX_BLUR)

instead of just PLATFORM(COCOA).

Fortunately this was noticed by Matt Woodrow as
fast/filter-image/clipped-filter.html started failing on iOS Simulator.
(Unfortunately not as part of the PR.)

(And this makes sense as HAVE(CGSTYLE_COLORMATRIX_BLUR) is not enabled
for iOS Simulator, though maybe it should be.)

Canonical link: https://commits.webkit.org/297046@main