elease Assert isInTopLayer() in Element::removeFromTopLayer() https://bugs.webkit.org/show_bug.cgi?id=295300 rdar://149686949
Reviewed by Ryosuke Niwa and Tim Nguyen.

The issue is caused by the race condition where the top layer is in
the process of being removed. The fix is to check the element it is
still the top layer before calling the removeFromTopLayer() function

MERGEING WITHOUT TEST REASON
============================
The test for this is not reliable. The test needs lots of reloads to
fail when it should (without the fix) and often timesout if the number
of reloads are more. The EWS failed with errors due to timeout or
texts being dumped when running EWS test. I am therefore merging
without tests

* Source/WebCore/html/HTMLElement.cpp:
(WebCore::HTMLElement::hidePopoverInternal):

Canonical link: https://commits.webkit.org/297056@main