[GTK][WPE] Use mpark::variant to reduce built library sizes https://bugs.webkit.org/show_bug.cgi?id=295461
Reviewed by Alex Christensen and Patrick Griffis.

Enable mpark::variant for the GTK and WPE ports. For example, library
sizes before and after for a WPE release build on x86_64 are:

 - Unstripped: 176373.2 KiB -> 168988.1 KiB (reduction: 7385.1 KiB)
 - Stripped..: 132520.7 KiB -> 131912.8 KiB (reduction:  607.9 KiB)

On top of the preprocessor guards, this needed a small change in
Source/WTF/wtf/Variant.h: be explicit with parens in in assertion
expression, and use WTF_ALLOW_UNSAFE_BUFFER_USAGE_{BEGIN,END} to avoid
GCC complaining. The rest of the changes take care of replacing
remaining std::variant uses with the WTF::Variant type alias.

Canonical link: https://commits.webkit.org/297077@main