[WebXR][OpenXR] XR_ERROR_SESSION_NOT_RUNNING issued sometimes when starting XR sessions https://bugs.webkit.org/show_bug.cgi?id=295526
Reviewed by Carlos Garcia Campos.

Added a boolean attribute that tracks the readiness status of a session.
It's true only after the session enters the READY state, and becomes
false after moving to STOPPING state (it's also the default value).

This way we wouldn't run xrWaitFrame->xrBeginFrame->xrEndFrame sequence
on sessions that are not ready. That could happen if WebXR is able to set
a callback for onFrameUpdate to the UI process before the OpenXR runtime
is able to begin the session.

* Source/WebKit/UIProcess/XR/openxr/PlatformXROpenXR.cpp:
(WebKit::OpenXRCoordinator::handleSessionStateChange):
(WebKit::OpenXRCoordinator::renderLoop):
* Source/WebKit/UIProcess/XR/openxr/PlatformXROpenXR.h:

Canonical link: https://commits.webkit.org/297102@main