[WebXR][OpenXR] Make graphics binding creation more secure https://bugs.webkit.org/show_bug.cgi?id=295517
Reviewed by Carlos Garcia Campos.

Added null pointer checks to the initialization of the graphics binding
for OpenXR. Also prepended "try" to the method to emphasize that it can
fail.

Took the chance to do some extra cleanups like the removal of unneeded
includes or replacing the surfaceless type for the more generic PlatformDisplay.

This is a followup of main@12938589

* Source/WebKit/UIProcess/XR/openxr/PlatformXROpenXR.cpp:
(WebKit::OpenXRCoordinator::tryInitializeGraphicsBinding):
(WebKit::OpenXRCoordinator::createSessionIfNeeded):
(WebKit::OpenXRCoordinator::initializeGraphicsBinding): Deleted.
* Source/WebKit/UIProcess/XR/openxr/PlatformXROpenXR.h:

Canonical link: https://commits.webkit.org/297103@main